backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
mysqlclient
django-cors-headers
gunicorn
python-dotenv